1、异常添加断电续传 2、修改停止后台服务方法，防止异常重启

+++ downloadlib/src/main/java/com/android/downloadlib/processor/entiry/ZDownloadBean.java
@@ -6,7 +6,7 @@ package com.android.downloadlib.processor.entiry;
  */
 
 public class ZDownloadBean {
-    public int progress;
+    public float progress;
     public long totalSize;
     public long downloadSize;
     public String speed;
